Code Projects/Project 21/Market Research/relevance_workforce/requirements.txt
python-dotenv>=1.0.1
relevanceai>=10.1.0
requests>=2.32.3
beautifulsoup4>=4.12.3
